moves Weather class to its own file
