+++ challenges/challenge-weather-app/script.js
@@ -1,3 +1,5 @@
+import { Weather } from "./weather.js";
+
 const thumbs = document.querySelector("#thumbs");
 const mainPhoto = document.querySelector("#photo");
 const conditions = document.querySelector("#conditions");
@@ -23,27 +25,6 @@ class EmptyOrNumericCity {
   }
 }
 const emptyOrNumericCity = new EmptyOrNumericCity();
-
-//==========Weather Class======================
-class Weather {
-  constructor(isFetching, city = "") {
-    this.isFetching = isFetching;
-    this.weatherData = {};
-    this.city = city;
-    emptyOrNumericCity.handleEmptyOrNumericCity(this.city);
-  }
-
-  static weatherAPIKey = config.weather_API_Key;
-
-  async fetchWeatherData() {
-    const url = `http://api.openweathermap.org/data/2.5/weather?q=${this.city}&appid=${Weather.weatherAPIKey}`;
-    const response = await fetch(url);
-    if (!response.ok) {
-      console.error(`Response status: ${response.status}`);
-    }
-    this.weatherData = await response.json();
-  }
-}
 
 //==========Photos Class======================
 class Photos {
